Conexion BackEnd con FrontEnd

+++ src/main/java/com/example/seguros/Service/ClienteService.java
@@ -1,6 +1,5 @@
 package com.example.seguros.Service;
 
-
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,26 +18,24 @@ public class ClienteService {
         this.clienteRepository = clienteRepository;
     }
 
-    public List<Cliente> guardarCliente(List<Cliente> cliente) { 
-    return clienteRepository.saveAll(cliente);
+    public List<Cliente> guardarCliente(List<Cliente> cliente) {
+        return clienteRepository.saveAll(cliente);
     }
 
-    public List<Cliente> obtenerTodos(){
+    public List<Cliente> obtenerTodos() {
         return clienteRepository.findAll();
     }
 
-    public Cliente obtenerPorDniCuit(String dniCuit){
+    public Cliente obtenerPorDniCuit(String dniCuit) {
         return clienteRepository.findByDniCuit(dniCuit)
-            .orElseThrow(() -> new IllegalArgumentException("Cliente no encontrado con dni: " + dniCuit));
+                .orElseThrow(() -> new IllegalArgumentException("Cliente no encontrado con dni: " + dniCuit));
     }
 
     public Cliente actualizarCliente(String dniCuit, Cliente clienteActualizado) {
         Cliente clienteExistente = obtenerPorDniCuit(dniCuit);
-        
-       
+
         clienteExistente.setNombre(clienteActualizado.getNombre());
         clienteExistente.setApellido(clienteActualizado.getApellido());
-        clienteExistente.setContraseña(clienteActualizado.getContraseña());
         clienteExistente.setEmail(clienteActualizado.getEmail());
         clienteExistente.setDireccion(clienteActualizado.getDireccion());
         clienteExistente.setTelefono(clienteActualizado.getTelefono());
@@ -46,8 +43,15 @@ public class ClienteService {
         return clienteRepository.save(clienteExistente);
     }
 
-    public void eliminarCliente(String dniCuit) {
-        Cliente cliente = obtenerPorDniCuit(dniCuit);
+    public void eliminarClientePorDni(String dniCuit) {
+        Cliente cliente = clienteRepository.findByDniCuit(dniCuit)
+                .orElseThrow(() -> new RuntimeException("Cliente no encontrado"));
+
+        // Validación: si el cliente tiene pólizas, no se puede borrar.
+        if (!cliente.getPolizas().isEmpty()) {
+            throw new RuntimeException("No se puede eliminar: el cliente tiene pólizas asociadas.");
+        }
+
         clienteRepository.delete(cliente);
     }
-}    
+}
